refactor: navigationBar 반응형 디자인 scss 적용 및 관련 html 코드 제거
