--- src/components/common/navigationBar/NavigationBar.module.scss
@@ -1,93 +1,0 @@
-@import '/src/styles/global.scss';
-
-.navBar {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: auto;
-  padding: 2rem 2rem 1rem 2rem;
-  gap: 1rem;
-
-  @include tablet {
-    align-items: center;
-    justify-content: space-between;
-  }
-}
-
-.logoButton_container {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  max-width: 96.4rem;
-
-  @include tablet {
-    flex-direction: row;
-    width: 100%;
-  }
-}
-
-.logo_container {
-  width: 8.4rem;
-  height: 3rem;
-}
-
-.searchBar-container-mobile {
-  position: relative;
-  width: 100%;
-
-  @include tablet {
-    display: none;
-  }
-}
-
-.searchBar-container-tablet {
-  display: none;
-  position: relative; /* 아이콘을 절대 위치로 설정할 수 있게 함 */
-
-  @include tablet {
-    display: flex;
-    flex-grow: 1;
-    margin: 0 2rem;
-  }
-}
-
-.buttons-container {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  font-size: 1.4rem;
-  font-weight: 700;
-}
-
-.notification {
-  display: flex;
-}
-
-.searchBar {
-  width: 100%;
-  height: 3.6rem;
-  padding-left: 4.2rem;
-  padding-right: 1.6rem;
-  border-radius: 1rem;
-  background-color: $color-gray10;
-
-  @include tablet {
-    margin-left: 4rem;
-  }
-}
-
-.searchBar::placeholder {
-  color: $color-gray40;
-}
-
-.searchBarIcon {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  left: 1.6rem;
-
-  @include tablet {
-    left: 5.6rem;
-  }
-}
