+++ src/layouts/feature-layout/Layout.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames/bind';
 import { ReactNode, RefObject } from 'react';
 import { Footer } from '@/src/components/common/ui-footer/Footer';
 import styles from './Layout.module.scss';
-import NavigationBar from '@/src/components/common/navigationBar/NavigationBar';
+import NavigationBar from '@/src/components/common/feature-navigation-bar/NavigationBar';
 
 const cx = classNames.bind(styles);
